// src/typings/Jitsi/JitsiConnectionEvents.d.ts
export const enum JitsiConnectionEvents {
  CONNECTION_DISCONNECTED = "connection.connectionDisconnected",
  CONNECTION_ESTABLISHED = "connection.connectionEstablished",
  CONNECTION_FAILED = "connection.connectionFailed",
  WRONG_STATE = "connection.wrongState",
  DISPLAY_NAME_REQUIRED = "connection.display_name_required"
}