// src/typings/Jitsi/JitsiTrackEvents.d.ts
export const enum JitsiTrackEvents {
    LOCAL_TRACK_STOPPED = "track.stopped",
    TRACK_AUDIO_LEVEL_CHANGED = "track.audioLevelsChanged",
    TRACK_AUDIO_OUTPUT_CHANGED = "track.audioOutputChanged",
    TRACK_MUTE_CHANGED = "track.trackMuteChanged",
    TRACK_VIDEOTYPE_CHANGED = "track.videoTypeChanged",
    NO_DATA_FROM_SOURCE = "track.no_data_from_source",
    NO_AUDIO_INPUT = "track.no_audio_input"
}